coverage: Deal with unused functions and their names in one place
